feat: implement multiplayer GPI with viewport-based player tracking
Replace Scene with Viewport that owns per-player state (local, activity,
region hash) via PlayerState array. Implement GPI protocol encoder with
4-section bit-packed updates, trait-based BitEncoders (Skip, LocalUpdate,
PlayerAdd, RegionUpdate), and recursive region hash delta for stale caches.

Consolidate player types to usize IDs, add PlayerSnapshot with mask flags,
extract InboxExt to codec, separate message processing from tick loop,
and fix bit writer mask calculation.

+++ net/src/message/login.rs
@@ -41,7 +41,7 @@ pub enum LoginOutcome {
 #[derive(Debug)]
 pub struct LoginSuccess {
     pub rights: u8,
-    pub player_index: u16,
+    pub player_index: usize,
     pub members: bool,
     pub inbox_tx: mpsc::Sender<Frame>,
     pub outbound_rx: mpsc::Receiver<Frame>,
@@ -100,7 +100,7 @@ impl LoginSuccess {
         buf.put_u8(0);
         buf.put_u8(0);
         buf.put_u8(0);
-        buf.put_u16(self.player_index);
+        buf.put_u16(self.player_index as u16);
         buf.put_u8(1);
         buf.put_u8(if self.members { 1 } else { 0 });
